+++ src/main/java/com/vdab/rdcar/controllers/EmployeeController.java
@@ -1,6 +1,6 @@
 package com.vdab.rdcar.controllers;
+
 import com.vdab.rdcar.domain.*;
-import com.vdab.rdcar.repositories.AppointmentRepository;
 import com.vdab.rdcar.services.AppointmentService;
 import com.vdab.rdcar.services.CarService;
 import com.vdab.rdcar.services.EmployeeService;
@@ -11,8 +11,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-
-import java.util.List;
+import static java.lang.Integer.parseInt;
 
 @Controller
 public class EmployeeController {
@@ -33,49 +32,49 @@ public class EmployeeController {
         return "index";
     }
 
-    @GetMapping(value= "/removeEmployee/{id}")
+    @GetMapping(value = "/removeEmployee/{id}")
     public String deleteAnEmployee(@PathVariable("id") Long id) {
+        Employee employee = employeeService.findById(id);
+        Integer maxKm = parseInt(employee.getCurrentCar().getMaxKm());
+        Integer currentMileage = parseInt(employee.getCurrentCarMileage());
+        if (currentMileage >= maxKm) {
             carService.deleteLease(carService.getLeasedByEmployeeId(id).getId());
-            employeeService.removeById(id);
+        }else{
+            LeasedCar lease = carService.getLeasedByEmployeeId(employee.getId());
+            lease.setEmployee(null);
+            carService.updateLease(lease);
+        }
+        employeeService.removeById(id);
 
         return "redirect:/";
     }
 
     @GetMapping(value = "/editEmployee/{id}")
-    public String showEditPage(@PathVariable("id") Long id, Model model){
-        model.addAttribute("editEmployee",employeeService.findById(id));
-        model.addAttribute("functionLevels" ,  FunctionLevels.values());
-        model.addAttribute("newEmployee" , new Employee());
+    public String showEditPage(@PathVariable("id") Long id, Model model) {
+        model.addAttribute("editEmployee", employeeService.findById(id));
+        model.addAttribute("functionLevels", FunctionLevels.values());
+        model.addAttribute("newEmployee", new Employee());
         return "edit";
     }
 
     @PostMapping(value = "/edit/{id}")
-    public String editEmployee(@ModelAttribute Employee newEmployee){
+    public String editEmployee(@ModelAttribute Employee newEmployee) {
         employeeService.updateEmployee(newEmployee);
         return "redirect:/";
     }
 
 
     @PostMapping(value = "/makeAppointment")
-    public String editAppointment(@ModelAttribute Appointment appointment){
+    public String editAppointment(@ModelAttribute Appointment appointment) {
         appointmentService.makeAppointment(appointment);
         return "redirect:/";
     }
 
 
-
-
     @GetMapping(value = "/addEmployee")
-    public String showAddPage(Model model){
-        model.addAttribute("newEmployee" , new Employee());
-        model.addAttribute("functionLevels" ,  FunctionLevels.values());
+    public String showAddPage(Model model) {
+        model.addAttribute("newEmployee", new Employee());
+        model.addAttribute("functionLevels", FunctionLevels.values());
         return "addEmployee";
     }
-
-    @PostMapping(value = "/add")
-    public String add(@ModelAttribute Employee newEmployee) {
-        employeeService.addEmployee(newEmployee);
-        return "redirect:/";
-    }
-
 }
